Complete Question 1
Completed the readItem function in C++ and the read_item function in Rust.
In both implementations, readItem and read_item take in an entire line as
a parameter and parses it to produce a GroceryItem.

If the C++ version fails, it throws an appropriate ParseError exception,
while the Rust version uses a variant type defined as follows:
- Result<GroceryItem, ParseError>
- On Success: Ok(record)
- On Failure: Err(error)